Refactor: Move NewType semantic analysis to a new module
